\name{mqext-package}
\alias{mqext-package}
\alias{mqext}
\docType{package}
\title{
M-quantile extention of the mixed linear model

}
\description{
M-quantile extension of the mixed linear model
}
\details{
\tabular{ll}{
Package: \tab mqext\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-11-26\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}


}
\author{
Tobias Schoch

Maintainer: Tobias Schoch <tobias.schoch@fhnw.ch>
}
\references{
Tzavidis et al. (2011)
}

\keyword{ package }
